refactor situational awareness
