+++ nebula/core/situationalawareness/candidateselection/hetcandidateselector.py
@@ -1,4 +1,4 @@
-from nebula.core.topologymanagement.candidateselection.candidateselector import CandidateSelector
+from nebula.core.situationalawareness.candidateselection.candidateselector import CandidateSelector
 from nebula.core.utils.locker import Locker
 
 class HETCandidateSelector(CandidateSelector):
