+++ nebula/core/situationalawareness/modelhandlers/aggmodelhandler.py
@@ -1,4 +1,4 @@
-from nebula.core.topologymanagement.modelhandlers.modelhandler import ModelHandler
+from nebula.core.situationalawareness.modelhandlers.modelhandler import ModelHandler
 from nebula.core.utils.locker import Locker
 
 class AGGModelHandler(ModelHandler):
