+++ nebula/core/situationalawareness/neighborpolicies/ringneighborpolicy.py
@@ -1,4 +1,4 @@
-from nebula.core.topologymanagement.neighborpolicies.neighborpolicy import NeighborPolicy
+from nebula.core.situationalawareness.neighborpolicies.neighborpolicy import NeighborPolicy
 from nebula.core.utils.locker import Locker
 import random
